+++ packages/debug_kit/lib/src/utils/export/debug_log_file_exporter.dart
@@ -2,10 +2,16 @@ import 'dart:io';
 import 'package:path_provider/path_provider.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:flutter/services.dart';
+import 'package:intl/intl.dart';
 import '../../core/models/debug_log_entry.dart';
 import 'debug_log_export_formatter.dart';
 
 class DebugLogFileExporter {
+  static String buildFileName(DateTime timestamp) {
+    final formatted = DateFormat('yyyyMMdd-HHmmss').format(timestamp);
+    return 'debugkit-logs-$formatted.txt';
+  }
+
   static Future<void> exportToClipboard(List<DebugLogEntry> logs) async {
     final content = DebugLogExportFormatter.formatLogs(logs);
     await Clipboard.setData(ClipboardData(text: content));
@@ -14,8 +20,8 @@ class DebugLogFileExporter {
   static Future<void> shareLogs(List<DebugLogEntry> logs) async {
     final content = DebugLogExportFormatter.formatLogs(logs);
     final directory = await getTemporaryDirectory();
-    final timestamp = DateTime.now().millisecondsSinceEpoch;
-    final file = File('${directory.path}/debug_kit_logs_$timestamp.txt');
+    final fileName = buildFileName(DateTime.now());
+    final file = File('${directory.path}/$fileName');
 
     await file.writeAsString(content);